+++ .gitignore
@@ -1,1 +1,3 @@
+node_modules,
+기타시도.txt
 node_modules
