+++ index.html
@@ -4,11 +4,24 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Document</title>
-  <link rel="stylesheet" href="./main.css">
-  <link rel="stylesheet" href="./reset.css">
-  <script type="module" src="./index.js"></script>
+  <link rel="stylesheet" href="./css/reset.css">
+  <link rel="stylesheet" href="./css/main.css">
 </head>
 <body>
+  <header>
+    <span>SOLMI KOH</span>
+    <div class="menu-container">
+      <a href="/">
+        <span>Human Resources</span>
+      </a>
+      <a href="/">
+        <span>More Info</span>
+      </a>
+      <a href="/">
+        <span>Playground</span>
+      </a>
+    </div>
+  </header>
   <h1>인재관리</h1>
   <button class="btn__add">등록</button>
   <dialog class="modal__add-profile">
@@ -29,45 +42,11 @@
         </div>
     </form>
   </dialog>
-  <div class="container">
-    <!-- <a href="/"  class="item">
-      <div>
-        <img class="profile-image" src="./assets/images/human_icon.png" alt="">
-        <div class="text-container">
-          <h1>이름임</h1>
-          <span>역할</span>
-          <span>깃헙</span>
-          <span>이메일</span>
-        </div>
-      </div>
-    </a>
-    <a href="/"  class="item">
-      <div>
-        <img class="profile-image" src="./assets/images/human_icon.png" alt="">
-        <div class="text-container">
-          <h1>이름임</h1>
-          <span>역할</span>
-          <span>깃헙</span>
-          <span>이메일</span>
-        </div>
-      </div>
-    </a>
-    <a href="/"  class="item">
-      <div>
-        <img class="profile-image" src="./assets/images/human_icon.png" alt="">
-        <div class="text-container">
-          <h1>이름임</h1>
-          <span>역할</span>
-          <span>깃헙</span>
-          <span>이메일</span>
-        </div>
-      </div>
-    </a> -->
-  </div>
+  <div class="container"></div>
 
   <!-- <div>
     <img id="myimg" src="" alt="test">
   </div> -->
-
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
